# Fix the crash in the compilation script in sync mode - [NGC-110]