--- a/package.json
+++ b/package.json
@@ -63,7 +63,7 @@
   "devDependencies": {
     "@incubateur-ademe/nosgestesclimat-scripts": "^0.3.1",
     "@incubateur-ademe/publicodes-commun": "^0.1.2",
-    "@publicodes/tools": "^0.4.3",
+    "@publicodes/tools": "^0.4.4",
     "@types/glob": "^8.1.0",
     "ansi-colors": "^4.1.3",
     "cli-progress": "^3.11.2",