The problem is that parsed rules from the engine use by the `constantFolding` function was [modified in place](https://github.com/publicodes/tools/pull/31) and this was a problem when called multiple times (I don't know why honestly).

This fixes https://github.com/publicodes/tools/issues/30 and improves some rule optimisations btw.

> [!IMPORTANT] 
> Doing a deep copy of the parsed rules noticeably degrades compilation performance. Therefore, this **should be considered as a temporary fix** while waiting to improve `constantFolding`.
> 
> `yarn compiles:rules`: from ~28s to 36s
> `yarn compiles:rules -o FR -t fr`: from ~3.71s to ~4.33s